chore(project): Complete Dockerization and backend refactoring
