+++ .env.example
@@ -18,6 +18,9 @@ INFLUXDB_BUCKET=your_influxdb_bucket
 KAFKA_BOOTSTRAP_SERVERS=localhost:9092
 KAFKA_DVR_GROUP_ID=dvr-group
 
-# --- Model Paths (Optional, can be hardcoded if they don't change) ---
-# STATUS_MODEL_PATH=backend/compressor_status_prediction_model.onnx
-# DART_MODEL_PATH=backend/dart_model.onnx
+# --- Model Paths ---
+STATUS_MODEL_PATH=artifacts/compressor_status_prediction_model.onnx
+DART_MODEL_PATH=artifacts/dart_model.onnx
+VIBRATION_MODEL_PATH=artifacts/farid_kaki_vibration_transformer.onnx
+SCALER_MEAN_PATH=artifacts/scaler_mean.npy
+SCALER_SCALE_PATH=artifacts/scaler_scale.npy
