+++ artifacts/compressor_status_prediction_model.onnx
