+++ artifacts/farid_kaki_vibration_transformer.onnx
