+++ artifacts/isolation_forest_model.onnx
