+++ artifacts/scaler_scale.npy
