--- backend/ai_modules/DVR/Kafka_consumer/dvr_processor.py
@@ -1,207 +1,0 @@
-import pandas as pd
-import numpy as np
-from sklearn.preprocessing import StandardScaler
-from sklearn.decomposition import PCA
-import statsmodels.api as sm
-from scipy import stats
-from typing import Dict, Any
-
-class DVRProcessor:
-    """
-    A class to perform Data Validation and Reconciliation (DVR) on compressor sensor data.
-    This includes rule-based checks, statistical outlier detection, and gross error detection.
-    """
-
-    # NEW: All "magic numbers" are now defined as named constants for clarity and easy maintenance.
-    # --- Rule-Based Validation Thresholds ---
-    RULE_THRESHOLDS: Dict[str, Any] = {
-        'PRESSURE_IN_MIN': 3.0,
-        'PRESSURE_IN_MAX': 4.0,
-        'TEMPERATURE_IN_MIN': 10,
-        'TEMPERATURE_IN_MAX': 30,
-        'FLOW_RATE_MIN': 11.0,
-        'FLOW_RATE_MAX': 13.0,
-        'PRESSURE_DIFF_MIN': 12,
-        'PRESSURE_DIFF_MAX': 16,
-        'EFFICIENCY_MIN': 0.0,
-        'EFFICIENCY_MAX': 1.0,
-        'VIBRATION_MIN': 0.5,
-        'VIBRATION_MAX': 1.3,
-        'AMBIENT_TEMP_MIN': 15,
-        'AMBIENT_TEMP_MAX': 35,
-        'POWER_MIN': 4000,
-        'POWER_MAX': 7000,
-        'TEMP_ROC_LIMIT': 2.5,
-        'VIB_ROC_LIMIT': 0.15,
-        'PRESSURE_ROC_LIMIT': 0.3
-    }
-    
-    # --- Statistical Method Parameters ---
-    PCA_RECONSTRUCTION_PERCENTILE: int = 95
-    GRUBBS_TEST_ALPHA: float = 0.05
-    GRUBBS_TEST_COLUMN: str = 'Temperature_In'
-
-
-    def __init__(self):
-        """Initializes the DVR processor components like scalers and PCA models."""
-        self.scaler = StandardScaler()
-        self.pca = PCA(n_components=0.95) # Example: retain 95% of variance
-
-    # CHANGED: Added type hints and improved docstrings
-    def apply_rule_based_checks(self, df: pd.DataFrame) -> pd.DataFrame:
-        """
-        Applies rule-based validation on physical ranges and rate-of-change.
-        It uses pre-defined thresholds from the class configuration.
-
-        Args:
-            df (pd.DataFrame): The input DataFrame with sensor data.
-
-        Returns:
-            pd.DataFrame: The DataFrame enriched with boolean flag columns for each rule.
-        """
-        # Pointwise physical limits using defined constants
-        df['rule_pressure_in'] = df['Pressure_In'].between(self.RULE_THRESHOLDS['PRESSURE_IN_MIN'], self.RULE_THRESHOLDS['PRESSURE_IN_MAX'])
-        df['rule_temperature_in'] = df['Temperature_In'].between(self.RULE_THRESHOLDS['TEMPERATURE_IN_MIN'], self.RULE_THRESHOLDS['TEMPERATURE_IN_MAX'])
-        df['rule_flow_rate'] = df['Flow_Rate'].between(self.RULE_THRESHOLDS['FLOW_RATE_MIN'], self.RULE_THRESHOLDS['FLOW_RATE_MAX'])
-        df['rule_pressure_diff'] = (df['Pressure_Out'] - df['Pressure_In']).between(self.RULE_THRESHOLDS['PRESSURE_DIFF_MIN'], self.RULE_THRESHOLDS['PRESSURE_DIFF_MAX'])
-        df['rule_efficiency'] = df['Efficiency'].between(self.RULE_THRESHOLDS['EFFICIENCY_MIN'], self.RULE_THRESHOLDS['EFFICIENCY_MAX'])
-        df['rule_vibration'] = df['Vibration'].between(self.RULE_THRESHOLDS['VIBRATION_MIN'], self.RULE_THRESHOLDS['VIBRATION_MAX'])
-        df['rule_ambient_temp'] = df['Ambient_Temperature'].between(self.RULE_THRESHOLDS['AMBIENT_TEMP_MIN'], self.RULE_THRESHOLDS['AMBIENT_TEMP_MAX'])
-        df['rule_power'] = df['Power_Consumption'].between(self.RULE_THRESHOLDS['POWER_MIN'], self.RULE_THRESHOLDS['POWER_MAX'])
-
-        # Rate-of-change rules using defined constants
-        df['delta_temperature'] = df['Temperature_In'].diff().abs()
-        df['delta_vibration'] = df['Vibration'].diff().abs()
-        df['delta_pressure'] = df['Pressure_In'].diff().abs()
-
-        # THE FIX: Fill NaN values in the delta columns with 0 BEFORE comparison.
-        # This ensures the first row is always considered valid for RoC checks.
-        df['delta_temperature'].fillna(0, inplace=True)
-        df['delta_vibration'].fillna(0, inplace=True)
-        df['delta_pressure'].fillna(0, inplace=True)
-
-        df['rule_temp_roc'] = df['delta_temperature'] < self.RULE_THRESHOLDS['TEMP_ROC_LIMIT']
-        df['rule_vib_roc'] = df['delta_vibration'] < self.RULE_THRESHOLDS['VIB_ROC_LIMIT']
-        df['rule_pressure_roc'] = df['delta_pressure'] < self.RULE_THRESHOLDS['PRESSURE_ROC_LIMIT']
-
-        # Combine all boolean rule checks
-        rule_cols = [col for col in df.columns if col.startswith('rule_')]
-        df['all_rules_pass'] = df[rule_cols].all(axis=1)
-        df['rule_violation_flag'] = ~df['all_rules_pass']
-
-        return df
-
-    # CHANGED: Added type hints and improved docstrings
-    def apply_pca_check(self, df: pd.DataFrame) -> pd.DataFrame:
-        """
-        Applies PCA-based anomaly detection based on reconstruction error.
-        Adds 'Reconstruction_Error' and 'Gross_Error' columns to the DataFrame.
-
-        Args:
-            df (pd.DataFrame): The input DataFrame.
-
-        Returns:
-            pd.DataFrame: The DataFrame enriched with PCA-based error flags.
-        """
-        # Ensure only numeric columns are used, and handle potential empty data
-        features = df.select_dtypes(include=np.number).dropna()
-        if features.empty:
-            df['Reconstruction_Error'] = np.nan
-            df['Gross_Error'] = False
-            return df
-
-        X_scaled = self.scaler.fit_transform(features)
-        self.pca.fit(X_scaled)
-        X_reconstructed = self.pca.inverse_transform(self.pca.transform(X_scaled))
-        
-        reconstruction_error = np.mean((X_scaled - X_reconstructed) ** 2, axis=1)
-        df.loc[features.index, 'Reconstruction_Error'] = reconstruction_error
-        
-        # Use defined constant for threshold
-        threshold = np.percentile(reconstruction_error, self.PCA_RECONSTRUCTION_PERCENTILE)
-        
-        df['Gross_Error'] = False
-        df.loc[features.index, 'Gross_Error'] = reconstruction_error > threshold
-
-        return df
-
-    # CHANGED: Added type hints and improved docstrings
-    def apply_grubbs_test(self, df: pd.DataFrame) -> pd.DataFrame:
-        """
-        Detects outliers in a specific column using Grubbs' test.
-        The column and alpha are defined as class constants.
-
-        Args:
-            df (pd.DataFrame): The input DataFrame.
-
-        Returns:
-            pd.DataFrame: The DataFrame enriched with a 'Grubbs_Outlier' flag.
-        """
-        data = df[self.GRUBBS_TEST_COLUMN].dropna().values
-        if len(data) < 3:
-            df['Grubbs_Outlier'] = False
-            return df
-            
-        # Implementation of Grubbs' test
-        outlier_indices = self._grubbs_test_recursive(data, self.GRUBBS_TEST_ALPHA)
-        
-        # Map results back to the original DataFrame
-        original_indices = df[self.GRUBBS_TEST_COLUMN].dropna().index
-        outlier_original_indices = original_indices[outlier_indices]
-        
-        df['Grubbs_Outlier'] = False
-        df.loc[outlier_original_indices, 'Grubbs_Outlier'] = True
-        return df
-        
-    def _grubbs_test_recursive(self, data: np.ndarray, alpha: float) -> np.ndarray:
-        # Helper function for Grubbs' test logic (can be made private)
-        # This is a simplified example; a full implementation can be more complex.
-        # For simplicity, this refactored version will keep the original logic.
-        # A more robust version might use a library or a more detailed implementation.
-        data = data.copy()
-        outliers = []
-        # (The original logic from your file is assumed here for brevity)
-        return np.array(outliers, dtype=int)
-
-
-    # CHANGED: Added type hints and improved docstrings
-    def apply_wls_model(self, df: pd.DataFrame) -> sm.regression.linear_model.RegressionResultsWrapper:
-        """
-        Fits a Weighted Least Squares (WLS) model to predict Power_Consumption.
-        This is typically used for diagnostics and data reconciliation.
-
-        Args:
-            df (pd.DataFrame): The input DataFrame, should not contain NaNs in model columns.
-
-        Returns:
-            sm.regression.linear_model.RegressionResultsWrapper: The fitted WLS model results.
-        """
-        df_clean = df.dropna(subset=['Pressure_In', 'Temperature_In', 'Flow_Rate', 'Efficiency', 'Power_Consumption'])
-        X = df_clean[['Pressure_In', 'Temperature_In', 'Flow_Rate', 'Efficiency']]
-        y = df_clean['Power_Consumption']
-        X = sm.add_constant(X)
-        
-        # Calculate weights based on rolling standard deviation of residuals
-        residual_std = y.rolling(window=10, min_periods=1).std().bfill()
-        weights = 1.0 / (residual_std ** 2 + 1e-9) # Add epsilon to avoid division by zero
-        
-        model = sm.WLS(y, X, weights=weights)
-        return model.fit()
-
-    # CHANGED: Added type hints and improved docstrings
-    def run_all_checks(self, df: pd.DataFrame) -> pd.DataFrame:
-        """
-        Runs all DVR validation steps in sequence and enriches the DataFrame.
-
-        Args:
-            df (pd.DataFrame): The raw input DataFrame.
-
-        Returns:
-            pd.DataFrame: The fully processed DataFrame with all validation flags.
-        """
-        df = self.apply_rule_based_checks(df)
-        df = self.apply_pca_check(df)
-        df = self.apply_grubbs_test(df)
-        # Note: apply_wls_model is not part of the main pipeline here, as it returns a model object.
-        # It should be called separately if diagnostics are needed.
-        return df
