+++ scripts/generate_scaler.py
